chore(test): address greptile cleanup feedback
