+++ libs/portal/xtream/feature/src/test-setup.ts
@@ -1,21 +1,7 @@
 import { setupZoneTestEnv } from 'jest-preset-angular/setup-env/zone';
+import { installDuplicateVideoJsQualityLevelsWarnFilter } from '@iptvnator/shared/testing';
 
-const originalConsoleWarn = console.warn.bind(console);
-console.warn = (...args: unknown[]) => {
-    const message = args
-        .filter((arg): arg is string => typeof arg === 'string')
-        .join(' ');
-
-    if (isDuplicateVideoJsQualityLevelsWarning(message)) {
-        return;
-    }
-
-    originalConsoleWarn(...args);
-};
-
-function isDuplicateVideoJsQualityLevelsWarning(message: string): boolean {
-    return message.includes('A plugin named "qualityLevels" already exists.');
-}
+installDuplicateVideoJsQualityLevelsWarnFilter();
 
 setupZoneTestEnv({
     errorOnUnknownElements: true,
