+++ libs/shared/testing/src/index.ts
@@ -1,0 +1,1 @@
+export { installDuplicateVideoJsQualityLevelsWarnFilter } from './lib/console-warn-filters';
