+++ tsconfig.base.json
@@ -79,6 +79,7 @@
             "@iptvnator/shared/m3u-utils": [
                 "libs/shared/m3u-utils/src/index.ts"
             ],
+            "@iptvnator/shared/testing": ["libs/shared/testing/src/index.ts"],
             "@iptvnator/services": ["libs/services/src/index.ts"],
             "@iptvnator/shared/interfaces": [
                 "libs/shared/interfaces/src/index.ts"
